src: avoiding compiler warnings in node_revert.cc
PR-URL: https://github.com/nodejs/node-private/pull/21

+++ src/node_revert.cc
@@ -45,7 +45,7 @@ void Revert(const char* cve) {
 #define V(code, label, _)                                                     \
   do {                                                                        \
     if (strcmp(cve, label) == 0) {                                            \
-      Revert(REVERT_ ## code);                                                \
+      Revert(static_cast<unsigned int>(REVERT_ ## code));                     \
       return;                                                                 \
     }                                                                         \
   } while (0);
@@ -63,7 +63,7 @@ bool IsReverted(const char * cve) {
 #define V(code, label, _)                                                     \
   do {                                                                        \
     if (strcmp(cve, label) == 0)                                              \
-      return IsReverted(REVERT_ ## code);                                     \
+      return IsReverted(static_cast<unsigned int>(REVERT_ ## code));          \
   } while (0);
   REVERSIONS(V)
   return false;
